geninfo: Make geninfo work more reliably on MSYS
From: Martin Hopfeld <martin.hopfeld@sse-erfurt.de>

Using the lcov tools on Win7 with MSYS and MinGW 4.5.1/4.5.2
raised some issues for us:

geninfo created in the for one source file in the 'SF:' line
paths starting with a lowercase drive letter and sometimes
starting with uppercase drive letters.

This lead to inaccurate coverage results on the MSYS platform.

This patch fixes this issue.